+++ website/src/playground/workers/prettierWorker.ts
@@ -1,9 +1,17 @@
+import prettier, { Options as PrettierOptions } from "prettier";
 import {
-	formatWithPrettier,
-	isTypeScriptFilename,
-	isJSONFilename,
-} from "../utils";
-import { defaultPlaygroundState, PlaygroundSettings } from "../types";
+	IndentStyle,
+	QuoteStyle,
+	QuoteProperties,
+	TrailingComma,
+	Semicolons,
+	PrettierOutput,
+	defaultPlaygroundState,
+	PlaygroundSettings,
+} from "../types";
+// @ts-ignore
+import parserBabel from "prettier/esm/parser-babel";
+import { isTypeScriptFilename } from "../utils";
 
 let settings = defaultPlaygroundState.settings;
 
@@ -51,3 +59,73 @@ self.addEventListener("message", (e) => {
 			console.error(`Unknown message ${e.data.type}.`);
 	}
 });
+
+function formatWithPrettier(
+	code: string,
+	options: {
+		lineWidth: number;
+		indentStyle: IndentStyle;
+		indentWidth: number;
+		filepath: string;
+		quoteStyle: QuoteStyle;
+		quoteProperties: QuoteProperties;
+		trailingComma: TrailingComma;
+		semicolons: Semicolons;
+	},
+): PrettierOutput {
+	try {
+		const prettierOptions: PrettierOptions = {
+			useTabs: options.indentStyle === IndentStyle.Tab,
+			tabWidth: options.indentWidth,
+			printWidth: options.lineWidth,
+			filepath: options.filepath,
+			plugins: [parserBabel],
+			parser: getPrettierParser(options.filepath),
+			singleQuote: options.quoteStyle === QuoteStyle.Single,
+			quoteProps: options.quoteProperties,
+			trailingComma: options.trailingComma,
+			semi: options.semicolons === Semicolons.Always,
+		};
+
+		// @ts-ignore
+		const debug = prettier.__debug;
+		const document = debug.printToDoc(code, prettierOptions);
+
+		// formatDoc must be before printDocToString because printDocToString mutates the document and breaks the ir
+		const ir = debug.formatDoc(document, {
+			parser: "babel",
+			plugins: [parserBabel],
+		});
+
+		const formattedCode = debug.printDocToString(
+			document,
+			prettierOptions,
+		).formatted;
+
+		return {
+			type: "SUCCESS",
+			code: formattedCode,
+			ir,
+		};
+	} catch (err: unknown) {
+		if (err instanceof SyntaxError) {
+			return {
+				type: "ERROR",
+				stack: err.message,
+			};
+		} else {
+			return {
+				type: "ERROR",
+				stack: (err as Error).stack ?? "",
+			};
+		}
+	}
+}
+
+function getPrettierParser(filename: string): string {
+	if (isTypeScriptFilename(filename)) {
+		return "babel-ts";
+	} else {
+		return "babel";
+	}
+}
